src/main/java/br/com/robertocarneiro/desafio_tech/service/impl/SaleServiceImpl.java
package br.com.robertocarneiro.desafio_tech.service.impl;

import br.com.robertocarneiro.desafio_tech.dto.Sale;
import br.com.robertocarneiro.desafio_tech.service.SaleService;
import br.com.robertocarneiro.desafio_tech.transformer.impl.SaleTransformer;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class SaleServiceImpl implements SaleService {

    private final SaleTransformer saleTransformer;

    @Override
    public List<Sale> transformByLines(List<String> lines) {
        return saleTransformer.transformByLines(lines);
    }
}
